feat(benchmark): add evaluation framework for testing AI agent performance (#2)
This commit implements a comprehensive benchmark evaluation framework:

- Add task evaluator agent for automated task evaluation
- Add evaluate CLI command for agent-based task testing
- Improve JSON extraction with multiple fallback strategies
- Add retry logic and increase max_tokens for LLM agents
- Fix safe UTF-8 truncation in error messages
- Complete benchmark evaluation framework structure
- Improve task generation to produce terminal-bench style tasks
- Enhance ideator prompts to generate harder benchmark tasks
- Include benchmark evaluation report with findings

Technical improvements:
- Shared json_extraction utilities across agents
- Security and code quality fixes from validation
- Test outputs directory excluded from git

Evaluation Results:
- 30% success rate achieved (target: 30-40% for Hard tasks)
- 10 tasks evaluated with difficulty calibration
- Tasks validated against terminal-bench format

+++ src/agents/full_pipeline_orchestrator.rs
@@ -727,7 +727,7 @@ impl FullPipelineOrchestrator {
                 task.difficulty.level,
                 task.metadata.tags.clone(),
                 &task.problem_statement,
-                (task.hidden_solution.expected_time_seconds / 60) as u32,
+                task.hidden_solution.expected_time_seconds / 60,
                 vec![],
             );
